aws_auth/serializers.py
from rest_framework import serializers
from aws_auth.models import Book, UserFile

class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = ['id', 'title', 'author']

class UserFileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserFile
        fields = ['id', 'file_name', 'file']