requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
mock
cryptography
boto3
django-storages[s3]
